Un `jar = Jar` and split out body bits
